made modifictions to pick cards refill function.
Instead of using just one 'special' object with details about the cards returend to the pick pile, I've decided to send a copy of each card tracking object that is sent back to the pick card tracker, with details of its movement.

That way, there are multiple copies of the object in breadcrumb, and on undo the one with the highest index before the current breadcrumb object can be used to restore the details on the tracking object for each card returned to waste card tracker.

I also fixed another issue.  When the undo was executed for the multiple card transfer to pick pile, moving all cards back to waste pile, I had forgotten to display the top card of the waste array, so this was causing an error because if that card had returned to the waste pile from somewhere else on the previous move, there was no card to return to that place causing an error to be thrown.