more changes to undo function.
instead of pushing a re-written copy of pick card file when sending to waste pile, I create a new object and populate using altered details.

(* typo, I meant a copy of pick card object which was in breadcrumb array)..
An issue occurred where the original breadcrumb was being altered; I had altered it, and then sent a copy of it to the breadcrumb to represent the move from pick tracker to waste tracker, but overlooked the fact that the older breadcrumb had now been altered, and this affected the undo function when that particular breadcrumb object was being used as a reference for the undo card.

Issue is now fixed since, instead of altering the previous breadcrumb, I just create a new one and give correct details and then send to breadcrumb array.

Also noticed that on undoing waste card moves where the undo reverted to the exact state of the piles when waste card pile was empty and pick card pile was full, the images for pick pile and waste pile were not showing as intended.  On undo, the pick pile should have the 'pick_reset' image as source and the empty waste pile should have the 'no more cards' image as source - FIXED