used JS to dynamically change class on cards dropped, depending on whether they are dropped to foundation piles or normal drop piles and whether the drops originate in foundation piles or waste piles.
The desired result was for any card dropped to foundation piles would be displayed directly on top of the card it covers, so on the foundation piles, only the most recent card added can be seen, and on the drop piles the cards cascade downwards as normal.

I still need to modify the auto solve card elements with the correct class so that they have the same behaviour of covering the previous foundation card when they are auto-dropped using the solve function.