all of unddo transitions working.
both single card and multiple card undo functions are working.

There are still some slight issues with the occasional card that is dropped to foundation pile, but not getting the foundationCardEl class, which prevents the card cascading.

There appears to be a bug with the undo function also, particularly when the issue in the above paragraph appears; and after a certain amount of undo's after the above error, the undo stops working with an error message.

I've added a sound for when cards are dropped onto the foundation pile