significant improvement in card drop reliability...
Previously, dropping cards to both foundation and drop piles was tricky because to drop the card, it had to be over the intended pile, but NOT over any of the children.  So, if the intended drop card was dragged over a pile, but while over the pile, was then also dragged over a child, the card would not drop on mouseup because the event target was not the pile, but the card that the intended drop card was directly over when mouseup occurred.

I originally dealt with this by giving the piles a padding value, which left a significant amount of space between the sides of a pile and the cards inside of it, and then then the intended drop card was between  the pile edge and the edge of a card inside the pile, the intended drop card would drop on mouseup.

This was a tricky move to do because, although the space between pile and its cards was significant, it was still a limited size because making it too large would then cause the size sum of all the piles to be greater than the screen width, which is not how the game is intended to look.  So it took quite a bit of practice to be able to posision the intended drop card correctly in order for it to drop, and often, the drop would fail because it was not difficult to overshoot the drop card so that it ended up entering the boundaries of a card already in the pile.

This resulted in having to repeatedly attempt to drop a card sometimes, which is not an efficient way for the game to be played; cards should only fail to drop if they don't meet the number/suit requirements dictated by the last faceup card in the pile.

ISSUE SOLVED:

I've solved this issue by using an IF/ELSE condition at the beginning of the drop function. Since all cards have an id that includes the string '.png', I used the condition that if the event.target.id includes that string, then the cardParent should be the event.target.parentNode, otherwise the cardParent is the event.target.

If on mouseup the intended drop card is over a card belonging to the pile, then the cardParent, which we want to append the intended drop card to, the event target is the card, so the cardParent variable is given the value, event.target.parentNode, which is the parent of the card, i.e. the pile that the card sits in.

In order for the drop to happen, the id of cardParent must be the id of one of the piles, and so, with this adjustment, the id becomes, not the target id, which is the id of the drop card was sitting over before mouseup, but the target parent id. And the recognition of this id allows the card to drop.

On the other hand if on mouseup the event.target id doesn't include the string '.png' then the value assigned to the cardParent variable is the event.target itself which is the pile.

This has madea  significant difference in drop reliability and speed since, the tricky positioning of the intended drop card between the edge of the pile cards and the pile itself is no longer required; just drag the card above the end card of the pile and the card can be sitting over either one of the elements, that is, the pile, or the pile card; either way the cardParent will be the pile.

This has made game play significantly faster and easier.

SOME ISSUES:
I noticed an issue this may have caused.  When undoing moves, and then putting other cards over a card that has been undone; the latest cards sit beneath the undone card, they are correctly positioned as the last card in the pile, but are covered by the card which was undone.  I noticed that the cards on which this happen still have the transition and translate styles on them in the DOM, so I have a hunch this might be the cause of the issue; I need to investigate this.

Other than that the fix has made game play less frustrating since most intended card drops are successful; there are still a few failures but very few by comparison.