extended the minimum size of drop piles so that when the pile is empty, there is a larger area shown for a card to be dropped to the pile.
the area has the same width of a card but a height of a few cards cascaded.

Since drop pile cards have a negative margin top, the piles sit lower than their top cards, so when a pile is empty, attempting to drop a card at the same vertical position where the pile's first card would normally sit doesn't work because the pile doesn't begin at that spot; it sits at a distance below that is equivalent to the margin top value of drop cards.  So when the pile is empty I might consider giving the same margin top value and then restoring the default value for populated pile positions when cards are dropped on the empty pile.